:recycle: Rename EAI to stadtbezirksbudget
