+++ stadtbezirksbudget-eai/src/test/java/de/muenchen/stadtbezirksbudget/TestConstants.java
@@ -1,4 +1,4 @@
-package de.muenchen.refarch;
+package de.muenchen.stadtbezirksbudget;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
