+++ stadtbezirksbudget-eai/src/test/java/de/muenchen/stadtbezirksbudget/configuration/LogbackJsonLoggingTest.java
